requirements.txt
# Core dependencies
numpy>=1.19.0
opencv-python>=4.5.0
python-dotenv>=0.19.0

# Voice
SpeechRecognition>=3.8.1
PyAudio>=0.2.11
gTTS>=2.2.0
pygame>=2.0.0

# AI/ML
openai>=0.27.0
requests>=2.25.0

# Hardware
RPi.GPIO>=0.7.0
pyserial>=3.5
pyserial-asyncio>=0.6
vilib>=1.0.0

# Development
gpiozero>=1.6.2
pytest>=6.2.0
pytest-cov>=2.8.0
black>=21.5b2
flake8>=3.9.0
mypy>=0.812

# Robot specific
robot-hat>=0.1.0
picamera2>=0.3.0
adafruit-circuitpython-servokit>=1.3.0

# Home Control
pychromecast>=13.0.0
zeroconf>=0.38.0
